refactor: separate the injection in gradle build from the automatic gradle builder
