+++ dspot/src/test/java/eu/stamp_project/automaticbuilder/AutomaticBuilderFactoryTest.java
@@ -1,5 +1,6 @@
 package eu.stamp_project.automaticbuilder;
 
+import eu.stamp_project.automaticbuilder.gradle.GradleAutomaticBuilder;
 import eu.stamp_project.program.InputConfiguration;
 import org.junit.After;
 import org.junit.Before;
